Embed script generator in module; Remove script
Remove the script file that is distributed with pycommand. Instead embed
the code in the ``if __name__ == '__main__:'`` block of the module.

This allows to invoke the code generator with ``python -m pycommand``
which is much cleaner. The actual code generator is now invoked using
a -g or --generate flag. This is done to make it easier to add more
functionality in the future.

The templates are now embedded in the module as well.